[DDW-118] - Setup ga tracking
